Final commit

+++ src/app/Component/products/products.component.ts
@@ -120,16 +120,20 @@ export class ProductComponent implements OnInit {
 
   deleteProduct(product: Products) {
     if (product._id) {
-      this.http.delete(`http://localhost:5000/api/products/${product._id}`).subscribe({
-        next: () => {
-          this.products = this.products.filter(p => p._id !== product._id);
-        },
-        error: (err) => {
-          console.error('Error deleting product:', err);
-        }
-      });
+      const isConfirmed = window.confirm("Are you sure you want to delete this product?");
+      
+      if (isConfirmed) {
+        this.http.delete(`http://localhost:5000/api/products/${product._id}`).subscribe({
+          next: () => {
+            this.products = this.products.filter(p => p._id !== product._id);
+          },
+          error: (err) => {
+            console.error('Error deleting product:', err);
+          }
+        });
+      }
     }
-  }
+  }  
 
   resetForm() {
     this.productName = '';
